refactor filter as a utility function
